# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Book API
  version: 1.0.0
paths:
  /api/books:
    get:
      summary: Get all books
      responses:
        '200':
          description: A list of books
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created
  /api/books/{id}:
    put:
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book updated
  /api/books/{id}/checkout:
    post:
      summary: Check out a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book checked out
        '404':
          description: Book not found
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        available:
          type: boolean
openapi: 3.0.0
info:
  title: Book API
  version: 1.0.0
paths:
  /api/books:
    get:
      summary: Get all books
      responses:
        '200':
          description: A list of books
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created
  /api/books/{id}:
    put:
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book updated
  /api/books/{id}/checkout:
    post:
      summary: Check out a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book checked out
        '404':
          description: Book not found
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        available:
          type: boolean